services/poller_engine/requirements.txt
requests>=2.31.0
schedule>=1.2.0
pika>=1.3.2
psycopg2-binary>=2.9.9
pysnmp-lextudio>=6.1.1
